indicator_app/pyproject.toml
[tool.poetry]
name = "indicator-app"
version = "0.1.0"
description = ""
authors = ["KBM15 <keyblademaster_01@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
numpy = "1.26.4"
tradingcore = "^0.1.0"
pika = "^1.3.2"
requests = "^2.32.3"
python-decouple = "^3.8"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
